Simplified keepAlive
The server sends updates every second during playback, which we acknowledge. that means we only need to keep the connection alive when playback is paused.
Also, we don't need to keep track of progression over time on the video position, as we can just feed the server the position it just fed us every time.